# Multiple changes to improve noise performance