@linuxianer99 2020-08-24
@hpmax : so we merge this to master ?
This will be avilable with sensord-testing.
sensord will still be the old version ...