@hpmax 2020-08-24
Works for me.  I definitely recommend this as a beta test, since I just don't have enough testing to prove it out.